requirements.txt
praw
python-dotenv
wordcloud
matplotlib
fuzzywuzzy
python-Levenshtein
titlecase
pandas